requirements.txt
# Required Waggle Python modules
pywaggle

# Custom Python modules. Please add your requirements below.
numpy

aiofiles==0.8.0
anyio==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
fastavro==1.4.10
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
jsonschema==4.4.0
kafka-python==2.0.2
pyrsistent==0.18.1
python-schema-registry-client==2.2.2
rfc3986==1.5.0
sniffio==1.2.0
zstandard==0.17.0
crc32c
python-dateutil
DateTime
